create admins file in routes folder and added superadmin register route into index file
