+++ routes/admins.js
@@ -1,0 +1,1 @@
+
